Split ElementInspector into inspector, shape editors and track editors
Nineteen composables and 1,148 lines in one file, covering three separate jobs:
the inspector and the placement/size/corner editors every element has, the
editors specific to one element kind (polygon, path, image, text, paint, repeat),
and the keyframe track editors. One file each.

The three cross-file editors widen from private to internal; nothing else changes.
